DO NOT MERGE
Record the average time a message takes to pass through the router.
Report these times when the egress link is closed.
Example output:

  TestRouterA: link MyReceiver count=500000 rx=1.805 aq=26.977 fwd=0.967 uq=51.053 tx=1.216

TestRouterA is the router name.
MyReceiver is the name of the egress link.
count is the number of messages forwarded out this link.
rx is the average time between initial message creation and placing the message on the core action list for forwarding (usecs).
aq is the average time the message waited on the action list.
fwd is the average time the core took to forward the message to the undelivered list on the outgoing link.
uq is the average time the message waited on the undelivered list before it was transmitted.
tx is the average time sending the message out the link.

The timing code is conditional - you enable it by defining QD_MESSAGE_TIMING at compilation:

  $ cmake .. -DCMAKE_INSTALL_PREFIX=/opt/kgiusti -DCMAKE_C_FLAGS="-DQD_MESSAGE_TIMING=1"